fix: move constant in file
